requirements.txt
streamlit
llama-cpp-python
langchain
langchain-community
chromadb
sentence-transformers
pypdf
